simplify and update to MVP version
